pyproject.toml
[project]
name = "simple-poll"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "django-vite>=3.0.5",
    "inertia-django>=0.6.0",
    "django-allauth[socialaccount]>=65.11.2",
    "whitenoise>=6.8.2",
    "gunicorn>=23.0.0",
    "nanoid>=2.0.0",
    "django-ipware>=7.0.1",
]
